Modified int to Integer Object. CRUD of RadCenter Updated. getRadCenterByRadCenterId service added.

+++ src/main/java/com/kinsolutions/radcenter/dao/RadCenterDao.java
@@ -18,7 +18,7 @@ public class RadCenterDao {
 	@PersistenceContext	
 	private EntityManager entityManager;
 
-	public RadCenter getRadCenterByUserId(long userId) {
+	public RadCenter getRadCenterByUserId(Integer userId) {
 		List<RadCenter> radCenterList = null;
 		try {
 			String hqlQuery = "FROM RadCenter WHERE users.userId = ?";
